+++ hooks/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "hooks",
-  "version": "0.0.1",
-  "description": "npm modules",
+  "name": "bunju-summit-hooks",
+  "version": "0.0.2",
+  "description": "A collection of React hooks for Bunju Summit.",
   "type": "module",
   "types": "dist/index.d.ts",
   "main": "dist/index.js",
