+++ hooks/src/lib/hooks/CardExpiryDate/utils/index.ts
@@ -1,4 +1,4 @@
-import { ValidationResult } from "@types";
+import { ValidationResult } from "../../../types";
 
 export const validateCardExpiryDate = (
   expiryDate: string
