+++ hooks/src/lib/hooks/CardNumber/index.ts
@@ -1,28 +1,7 @@
 import { useState } from "react";
 import { INITIAL_CARD_NUMBER_STATE } from "./constants";
 import { CardNumberKey, CardNumberState } from "./types";
-import { validateCardNumber } from "./utils";
-import { ValidationResult } from "@types";
-
-const validateCardNumbers = (cardNumber: CardNumberState) => {
-  let errorState: ValidationResult = {
-    isValid: true,
-    errorMessage: "",
-  };
-
-  Object.values(cardNumber).every(({ value }) => {
-    console.log(cardNumber);
-
-    const { isValid, errorMessage } = validateCardNumber(value);
-    errorState = { isValid, errorMessage };
-
-    if (!isValid) {
-      return false;
-    }
-    return true;
-  });
-  return errorState;
-};
+import { validateCardNumbers } from "./utils";
 
 const useCardNumber = () => {
   const [cardNumberState, setCardNumberState] = useState<CardNumberState>(
